trying hacky solutions now
